split out s2_inference errors into its own file
