+++ src/marqo/errors.py
@@ -1,22 +1,6 @@
 import json
 from requests import Response
 from http import HTTPStatus
-from typing import Optional
-
-
-class S2InferenceError(Exception):
-    def __init__(self, message: Optional[str] = None) -> None:
-        if Optional is not None:
-            self.message = message
-            super().__init__(self.message)
-
-
-class ChunkerError(S2InferenceError):
-    pass
-
-
-class VectoriseError(S2InferenceError):
-    pass
 
 
 class MarqoError(Exception):
