+++ src/marqo/s2_inference/errors.py
@@ -1,0 +1,16 @@
+from typing import Optional
+
+
+class S2InferenceError(Exception):
+    def __init__(self, message: Optional[str] = None) -> None:
+        if Optional is not None:
+            self.message = message
+            super().__init__(self.message)
+
+
+class ChunkerError(S2InferenceError):
+    pass
+
+
+class VectoriseError(S2InferenceError):
+    pass
